Modified Kaprekar Numbers
Validated all test cases for Modified Kaprekar Numbers

+++ HackerRankChallenges/ModifiedKaprekarNumbers.cs
@@ -12,7 +12,7 @@ namespace HackerRankChallenges
 
             for (int i = startRange; i <= endRange; i++)
             {
-                ulong squareRootOfWholeNumber = (ulong)(i * i);
+                ulong squareRootOfWholeNumber = (ulong)i * (ulong)i;
 
                 var squareRootOfWholeNumberAsText = squareRootOfWholeNumber.ToString();
